More datatype corrections for curl
Most of these are just to silence gcc warnings but the part where we wrote
to a double as if it were a curl_off_t (a 64 bit integer type) and then put
that into a long long using assignment (and thus implicit conversion rules)
could've gone wrong in some cases.

+++ src/fetchmail.c
@@ -4558,7 +4558,7 @@ bool mailDescend(const char *title, char cmd)
 	struct FOLDER f0;
 
 	active_a = a, isimap = true;
-	curl_easy_setopt(h, CURLOPT_VERBOSE, (debugLevel >= 4));
+	curl_easy_setopt(h, CURLOPT_VERBOSE, (long) (debugLevel >= 4));
 	path = cw->baseDirName;
 	f0.path = path; // that's all we need in f0
